increased security on most functions + restricted functions for unregistered users

+++ src/Controller/PostController.php
@@ -6,6 +6,7 @@ use App\Entity\Post;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class PostController extends AbstractController
@@ -22,16 +23,18 @@ class PostController extends AbstractController
 
     /**
      * @Route("/post/{id}/delete", name="delete_post")
+     * @IsGranted("ROLE_USER")
      */
     public function delete(ManagerRegistry $doctrine, Post $post): Response
     {
-
+        
+        if($post->getAuteur()->getId() == $this->getUser()->getId()){
+            $entityManager = $doctrine->getManager();
+            $entityManager->remove($post);
+            $entityManager->flush();
+        }
+        
         $id = $post->getTopic()->getId();
-
-        $entityManager = $doctrine->getManager();
-        $entityManager->remove($post);
-        $entityManager->flush();
-
         return $this->redirectToRoute('show_topic', ['id' => $id]);
     }
